split the innodb queries on a different file
